+++ tests/main.rs
@@ -1815,6 +1815,8 @@ fn test_nft_sanity() {
 
     e.skip_time(10);
 
+    e.dao_add_approved_nft(&nfts.paras);
+
     e.nft_transfer(&users.charlie, &nfts.paras, &users.dude, &nft_id);
     let nft = e.get_nft_token(&nfts.paras, &nft_id).unwrap();
     assert_eq!(
@@ -1822,8 +1824,7 @@ fn test_nft_sanity() {
         String::from(&stream_id)
     );
 
-    // TODO enable #11
-    /*let stream = e.get_stream(&stream_id);
+    let stream = e.get_stream(&stream_id);
     assert_eq!(
         e.get_balance(&tokens.wnear_simple, &users.charlie),
         d(10, 23) / 250 * 249
@@ -1844,7 +1845,7 @@ fn test_nft_sanity() {
         e.get_balance(&tokens.wnear_simple, &users.dude),
         d(20, 23) / 250 * 249
     );
-    assert_eq!(stream.balance, d(70, 23) - d(1, 23));*/
+    assert_eq!(stream.balance, d(70, 23) - d(1, 23));
 
     e.nft_detach_stream(&nfts.paras, &nft_id, &stream_id);